move fs & s3 common code to abstract class; combine n5 readers/writers
