--- src/main/java/org/janelia/saalfeldlab/n5/s3/N5S3Writer.java
@@ -1,263 +1,0 @@
-/**
- * Copyright (c) 2017, Stephan Saalfeld
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- */
-package org.janelia.saalfeldlab.n5.s3;
-
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.DataOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.lang.reflect.Type;
-import java.nio.channels.Channels;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
-
-import org.janelia.saalfeldlab.n5.BlockWriter;
-import org.janelia.saalfeldlab.n5.CompressionType;
-import org.janelia.saalfeldlab.n5.DataBlock;
-import org.janelia.saalfeldlab.n5.DataType;
-import org.janelia.saalfeldlab.n5.DatasetAttributes;
-import org.janelia.saalfeldlab.n5.N5Writer;
-
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.DeleteObjectsRequest;
-import com.amazonaws.services.s3.model.ListObjectsV2Request;
-import com.amazonaws.services.s3.model.ListObjectsV2Result;
-import com.amazonaws.services.s3.model.ObjectMetadata;
-import com.amazonaws.services.s3.model.S3ObjectSummary;
-import com.google.gson.GsonBuilder;
-import com.google.gson.JsonElement;
-import com.google.gson.reflect.TypeToken;
-
-/**
- * A simple structured container format for hierarchies of chunked
- * n-dimensional datasets and attributes.
- *
- * {@linkplain https://github.com/axtimwalde/n5}
- *
- * @author Stephan Saalfeld
- */
-public class N5S3Writer extends N5S3Reader implements N5Writer {
-
-	/**
-	 * Opens an {@link N5Writer} at a given base path with a custom
-	 * {@link GsonBuilder} to support custom attributes.
-	 *
-	 * If the base path does not exist, it will not be created and all
-	 * subsequent attempts to read or write attributes, groups, or datasets
-	 * will fail with an {@link IOException}.
-	 *
-	 * If the base path is not writable, all subsequent attempts to write
-	 * attributes, groups, or datasets will fail with an {@link IOException}.
-	 *
-	 * @param basePath n5 base path
-	 * @param gsonBuilder
-	 */
-	public N5S3Writer(final AmazonS3 s3, final String bucket, final GsonBuilder gsonBuilder) {
-
-		super(s3, bucket, gsonBuilder);
-	}
-
-	/**
-	 * Opens an {@link N5Writer} at a given base path.
-	 *
-	 * If the base path does not exist, it will not be created and all
-	 * subsequent attempts to read or write attributes, groups, or datasets
-	 * will fail with an {@link IOException}.
-	 *
-	 * If the base path is not writable, all subsequent attempts to write
-	 * attributes, groups, or datasets will fail with an {@link IOException}.
-	 *
-	 * @param basePath n5 base path
-	 */
-	public N5S3Writer(final AmazonS3 s3, final String bucket) {
-
-		this(s3, bucket, new GsonBuilder());
-	}
-
-	@Override
-	public void createContainer() throws IOException {
-
-		if (!s3.doesBucketExistV2(bucket))
-			s3.createBucket(bucket);
-		createGroup("");
-	}
-
-	@Override
-	public void removeContainer() throws IOException {
-
-		remove("");
-		s3.deleteBucket(bucket);
-	}
-
-	@Override
-	public void setAttributes(final String pathName, final Map<String, ?> attributes) throws IOException {
-
-		final String metadataKey = Paths.get(pathName, jsonFile).toString();
-		final HashMap<String, JsonElement> map = getAttributes(pathName);
-		for (final Entry<String, ?> entry : attributes.entrySet())
-			map.put(entry.getKey(), gson.toJsonTree(entry.getValue()));
-
-		final Type mapType = new TypeToken<HashMap<String, JsonElement>>(){}.getType();
- 		final String json = gson.toJson(map, mapType);
-
-		final byte[] bytes = json.getBytes("UTF-8");
-		try (final InputStream data = new ByteArrayInputStream(bytes)) {
-			final ObjectMetadata objectMetadata = new ObjectMetadata();
-			objectMetadata.setContentLength(bytes.length);
-			s3.putObject(bucket, removeFrontDelimiter(metadataKey), data, objectMetadata);
-		}
-	}
-
-	@Override
-	public void createGroup(final String pathName) throws IOException {
-
-		final Path path = Paths.get(pathName);
-		for (int i = 0; i < path.getNameCount(); ++i) {
-			final String subgroup = path.subpath(0, i + 1).toString();
-			if (!exists(subgroup))
-				setAttributes(subgroup, Collections.emptyMap());
-		}
-	}
-
-	/**
-	 * Removes a group or dataset (directory and all contained files).
-	 *
-	 * <p><code>{@link #remove(String) remove("")}</code> or
-	 * <code>{@link #remove(String) remove("")}</code> will delete this N5
-	 * container.  Please note that no checks for safety will be performed,
-	 * e.g. <code>{@link #remove(String) remove("..")}</code> will try to
-	 * recursively delete the parent directory of this N5 container which
-	 * only fails because it attempts to delete the parent directory before it
-	 * is empty.
-	 *
-	 * @param pathName group path
-	 * @throws IOException
-	 */
-	@Override
-	public boolean remove(final String pathName) throws IOException {
-
-		final String correctedPathName = removeFrontDelimiter(pathName);
-		final String prefix = !correctedPathName.isEmpty() ? appendDelimiter(correctedPathName) : correctedPathName;
-		final ListObjectsV2Request listObjectsRequest = new ListObjectsV2Request()
-				.withBucketName(bucket)
-				.withPrefix(prefix);
-		ListObjectsV2Result objectsListing;
-		do {
-			objectsListing = s3.listObjectsV2(listObjectsRequest);
-			final List<String> objectsToDelete = new ArrayList<>();
-			for (final S3ObjectSummary object : objectsListing.getObjectSummaries())
-				objectsToDelete.add(object.getKey());
-
-			if (!objectsToDelete.isEmpty()) {
-				s3.deleteObjects(new DeleteObjectsRequest(bucket)
-						.withKeys(objectsToDelete.toArray(new String[objectsToDelete.size()]))
-					);
-			}
-			listObjectsRequest.setContinuationToken(objectsListing.getNextContinuationToken());
-		} while (objectsListing.isTruncated());
-		return !exists(pathName);
-	}
-
-	/**
-	 * Creates a dataset.  This does not create any data but the path and
-	 * mandatory attributes only.
-	 *
-	 * @param pathName dataset path
-	 * @param datasetAttributes
-	 * @throws IOException
-	 */
-	@Override
-	public void createDataset(
-			final String pathName,
-			final DatasetAttributes datasetAttributes) throws IOException{
-
-		createGroup(pathName);
-		setDatasetAttributes(pathName, datasetAttributes);
-	}
-
-	/**
-	 * Creates a dataset.  This does not create any data but the path and
-	 * mandatory attributes only.
-	 *
-	 * @param pathName dataset path
-	 * @param dimensions
-	 * @param blockSize
-	 * @param dataType
-	 * @throws IOException
-	 */
-	@Override
-	public void createDataset(
-			final String pathName,
-			final long[] dimensions,
-			final int[] blockSize,
-			final DataType dataType,
-			final CompressionType compressionType) throws IOException {
-
-		createGroup(pathName);
-		setDatasetAttributes(pathName, new DatasetAttributes(dimensions, blockSize, dataType, compressionType));
-	}
-
-	@Override
-	public <T> void writeBlock(
-			final String pathName,
-			final DatasetAttributes datasetAttributes,
-			final DataBlock<T> dataBlock) throws IOException {
-
-		final String dataBlockKey = getDataBlockPath(pathName, dataBlock.getGridPosition()).toString();
-		try (final ByteArrayOutputStream byteStream = new ByteArrayOutputStream()) {
-			final DataOutputStream dos = new DataOutputStream(byteStream);
-
-			final int mode = (dataBlock.getNumElements() == DataBlock.getNumElements(dataBlock.getSize())) ? 0 : 1;
-			dos.writeShort(mode);
-
-			dos.writeShort(datasetAttributes.getNumDimensions());
-			for (final int size : dataBlock.getSize())
-				dos.writeInt(size);
-
-			if (mode != 0)
-				dos.writeInt(dataBlock.getNumElements());
-
-			dos.flush();
-
-			final BlockWriter writer = datasetAttributes.getCompressionType().getWriter();
-			writer.write(dataBlock, Channels.newChannel(dos));
-
-			final byte[] bytes = byteStream.toByteArray();
-			try (final InputStream data = new ByteArrayInputStream(bytes)) {
-				final ObjectMetadata objectMetadata = new ObjectMetadata();
-				objectMetadata.setContentLength(bytes.length);
-				s3.putObject(bucket, removeFrontDelimiter(dataBlockKey), data, objectMetadata);
-			}
-		}
-	}
-}
